Update Pow Brew file to include SHA256 hash
Homebrew is deprecating all SHA1 hashes in favor of SHA256. The
version of the Pow Homebrew formula we had still had a SHA1 hash
which led to lots of deprication warnings and stuff. This updates the
formula to have a SHA256 hash. New formula from homebrew core here
(https://github.com/Homebrew/homebrew-core/blob/master/Formula/pow.rb)

It keeps the version number the same as it was before, since that has
a boxen post-fix.